requirements.txt
# Core (Apple Silicon / MPS compatible)
torch>=2.1.0
torchvision>=0.16.0
accelerate>=0.25.0
diffusers[training]>=0.24.0
transformers>=4.36.0
peft>=0.7.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Data / metrics
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
kagglehub>=0.2.0
safetensors>=0.4.0
# bitsandbytes — CUDA-only, no macOS arm64 wheel; removed
# xformers — CUDA-only; MPS uses torch SDPA natively; removed

# Visualization (optional, for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
